requirements.txt
python-dotenv
requests
slackclient